+++ jv_builder_base.py
@@ -1,0 +1,12 @@
+class JVBuilderBase:
+    @staticmethod
+    def draw(prop_group, layout):
+        pass
+
+    @staticmethod
+    def update(prop_group, obj):
+        pass
+
+    @staticmethod
+    def delete(prop_group, obj):
+        pass
