+++ jv_types.py
@@ -1,0 +1,10 @@
+from jv_flooring import JVFlooring
+
+
+registered_types = {
+    "flooring": JVFlooring
+}
+
+
+def get_object_type_handler(object_type: str):
+    return registered_types.get(object_type, None)
